Handle a specific case of anyOf
The `anyOf` handling was not considering a case that's already coded in
the kubenix generator, the int-or-string case.

For reasons beyond my current knowledge, the CRDs I'm using
(Cloudnative-PG) contain a property called `x-kubernetes-int-or-string`
and not the expected `int-or-string` in the `format` property inside the
`type`.

This commit handles this case by patching up the definition so that
the generator uses `either int str` instead of blindly picking the first
one.