fix: visible failure in github actions UI when tests fail
Note the documentation on [workflow exit
codes](https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#exit-codes-and-error-action-preference).
While the default shell for job steps is in fact bash, there is a
difference between the default and explicitly defining `shell: bash`. A
clearer description of this difference is
[here](https://stackoverflow.com/a/75420766).

Essentially, without `-o pipefail`, the exit code will be 0 (success),
even if the test fails, because we pipe the result of the testing
commands into `tee` which will always succeed. It isn't clear without
further investigation that something is actually wrong with the tests,
so setting `-o pipefail` with explicit `shell: bash` will propagate the
exit code of the actual testing command to ensure it reflects the actual
workflow result.