fix: allow _[a-zA-Z]+ string for an attr key.
### Problem
Reference the added test.

### Solution
Only exclude `_priority` instead of `_[a-zA-Z0-9_./]*`.

While the original intention of this exclusion isn't clear from the kubenix origin commit [^1], 
Kubernetes allows `[a-zA-Z_.-/][a-zA-Z0-9_.-/]*` for the property names
with a few exeception rules, none of which prohibits the underscore
prefix regex `_[a-zA-Z0-9.-/]*` [^2].

I infer the purpose being a future proof for some nix-specific internal option fields
such as `_priority`. However, as of v1.33.0 k8s version, only `_priority` is defined by the nix generator [^3].
Now that nixidy generates the k8s modules [^4], any additional internal
options prefixed with `_` in the future can be coordinated.

[^1]: kubenix first commit: https://github.com/hall/kubenix/commit/cbf84e25a574b9bad1a97e1e29a2d5b2fe7d660a
[^2]: https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definitions/#validation
[^3]: k8s jsonschema doesn't define _priority option: https://github.com/search?q=repo%3Ayannh%2Fkubernetes-json-schema+path%3A%2F%5Ev1%5C.33%5C.0%5C%2F%2F++_priority&type=code 
[^4]: https://github.com/arnarg/nixidy/blob/main/docs/developer_guide/architecture.md?plain=1#L82